threshold change

+++ data/multimercluster_fast.sh
@@ -8,45 +8,55 @@ notExists() {
     [ ! -f "$1" ]
 }
 
-if notExists "${TMP_PATH}/chain_clu.dbtype"; then
-    # shellcheck disable=SC2086
-    "$MMSEQS" cluster "${INPUT}" "${TMP_PATH}/chain_clu" "${TMP_PATH}/chaincluster_tmp" --remove-tmp-files 0 --tmscore-threshold "${CHAIN_TM_THRESHOLD}" -c "${CHAIN_CLUSTER_C}" --cov-mode "${CHAIN_CLUSTER_COV_MODE}" ${CHAINCLUSTER_PAR} \
-        || fail "chain cluster died"
-fi
+if [ -n "${USE_ORIGINAL_MULTIMERCLUSTER}" ]; then
+    if notExists "${RESULT}.dbtype"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" multimercluster "${INPUT}" "${RESULT}" "${TMP_PATH}" ${MULTIMERCLUSTER_PAR} \
+            || fail "multimercluster died"
+    fi
+else
+    if notExists "${TMP_PATH}/chain_clu.dbtype"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" cluster "${INPUT}" "${TMP_PATH}/chain_clu" "${TMP_PATH}/chaincluster_tmp" --remove-tmp-files 0 --tmscore-threshold "${CHAIN_TM_THRESHOLD}" -c "${CHAIN_CLUSTER_C}" --cov-mode "${CHAIN_CLUSTER_COV_MODE}" ${CHAINCLUSTER_PAR} \
+            || fail "chain cluster died"
+    fi
 
-if notExists "${TMP_PATH}/candidate_pref.dbtype"; then
-    # shellcheck disable=SC2086
-    "$MMSEQS" chainmultimerprefilter "${INPUT}" "${INPUT}" "${TMP_PATH}/chain_clu" "${TMP_PATH}/candidate_pref" ${CHAINMULTIMERPREFILTER_PAR} \
-        || fail "chainmultimerprefilter died"
-fi
+    if notExists "${TMP_PATH}/candidate_pref.dbtype"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" chainmultimerprefilter "${INPUT}" "${INPUT}" "${TMP_PATH}/chain_clu" "${TMP_PATH}/candidate_pref" ${CHAINMULTIMERPREFILTER_PAR} \
+            || fail "chainmultimerprefilter died"
+    fi
 
-if notExists "${TMP_PATH}/candidate_aln.dbtype"; then
-    # shellcheck disable=SC2086
-    "$MMSEQS" structurealign "${INPUT}" "${INPUT}" "${TMP_PATH}/candidate_pref" "${TMP_PATH}/candidate_aln" ${STRUCTUREALIGN_PAR} \
-        || fail "structurealign died"
-fi
+    if notExists "${TMP_PATH}/candidate_aln.dbtype"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" structurealign "${INPUT}" "${INPUT}" "${TMP_PATH}/candidate_pref" "${TMP_PATH}/candidate_aln" ${STRUCTUREALIGN_PAR} \
+            || fail "structurealign died"
+    fi
 
-if notExists "${TMP_PATH}/multimer_result.dbtype"; then
-    # shellcheck disable=SC2086
-    "$MMSEQS" scoremultimer "${INPUT}" "${INPUT}" "${TMP_PATH}/candidate_aln" "${TMP_PATH}/multimer_result" ${SCOREMULTIMER_PAR} \
-        || fail "scoremultimer died"
-fi
+    if notExists "${TMP_PATH}/multimer_result.dbtype"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" scoremultimer "${INPUT}" "${INPUT}" "${TMP_PATH}/candidate_aln" "${TMP_PATH}/multimer_result" ${SCOREMULTIMER_PAR} \
+            || fail "scoremultimer died"
+    fi
 
-if notExists "${RESULT}.dbtype"; then
-    # shellcheck disable=SC2086
-    "$MMSEQS" clust "${INPUT}" "${TMP_PATH}/multimer_result" "${RESULT}" ${CLUSTER_PAR} \
-        || fail "Clustering died"
+    if notExists "${RESULT}.dbtype"; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" clust "${INPUT}" "${TMP_PATH}/multimer_result" "${RESULT}" ${CLUSTER_PAR} \
+            || fail "Clustering died"
+    fi
 fi
 
 if [ -n "${REMOVE_TMP}" ]; then
-    # shellcheck disable=SC2086
-    "$MMSEQS" rmdb "${TMP_PATH}/multimer_result" ${VERBOSITY_PAR} || fail "rmdb died"
-    # shellcheck disable=SC2086
-    "$MMSEQS" rmdb "${TMP_PATH}/candidate_pref" ${VERBOSITY_PAR} || fail "rmdb died"
-    # shellcheck disable=SC2086
-    "$MMSEQS" rmdb "${TMP_PATH}/candidate_aln" ${VERBOSITY_PAR} || fail "rmdb died"
-    # shellcheck disable=SC2086
-    "$MMSEQS" rmdb "${TMP_PATH}/chain_clu" ${VERBOSITY_PAR} || fail "rmdb died"
-    rm -rf -- "${TMP_PATH}/chaincluster_tmp"
+    if [ -z "${USE_ORIGINAL_MULTIMERCLUSTER}" ]; then
+        # shellcheck disable=SC2086
+        "$MMSEQS" rmdb "${TMP_PATH}/multimer_result" ${VERBOSITY_PAR} || fail "rmdb died"
+        # shellcheck disable=SC2086
+        "$MMSEQS" rmdb "${TMP_PATH}/candidate_pref" ${VERBOSITY_PAR} || fail "rmdb died"
+        # shellcheck disable=SC2086
+        "$MMSEQS" rmdb "${TMP_PATH}/candidate_aln" ${VERBOSITY_PAR} || fail "rmdb died"
+        # shellcheck disable=SC2086
+        "$MMSEQS" rmdb "${TMP_PATH}/chain_clu" ${VERBOSITY_PAR} || fail "rmdb died"
+        rm -rf -- "${TMP_PATH}/chaincluster_tmp"
+    fi
     rm -f -- "${TMP_PATH}/multimercluster_fast.sh"
 fi
